+++ SpringBoot/src/main/java/com/seungwook/ktsp/domain/board/type/community/notice/controller/NoticeCommandController.java
@@ -2,7 +2,7 @@ package com.seungwook.ktsp.domain.board.type.community.notice.controller;
 
 import com.seungwook.ktsp.domain.board.type.community.common.dto.request.CommunityRegisterRequest;
 import com.seungwook.ktsp.domain.board.type.community.common.dto.request.CommunityUpdateRequest;
-import com.seungwook.ktsp.domain.board.type.community.notice.service.NoticeService;
+import com.seungwook.ktsp.domain.board.type.community.notice.service.NoticeCommandService;
 import com.seungwook.ktsp.global.auth.support.AuthHandler;
 import com.seungwook.ktsp.global.response.Response;
 import io.swagger.v3.oas.annotations.Operation;
@@ -18,13 +18,13 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/admin/notice")
 public class NoticeCommandController {
 
-    private final NoticeService noticeService;
+    private final NoticeCommandService noticeCommandService;
 
     @Operation(summary = "공지사항 등록", description = "공지사항은 Admin만 등록 가능")
     @PostMapping
     public ResponseEntity<Response<Void>> registerNotice(@Valid @RequestBody CommunityRegisterRequest request) {
 
-        noticeService.registerNotice(AuthHandler.getUserId(), request);
+        noticeCommandService.registerNotice(AuthHandler.getUserId(), request);
 
         return ResponseEntity.ok(Response.<Void>builder()
                 .message("공지사항 등록 성공")
@@ -35,7 +35,7 @@ public class NoticeCommandController {
     @PatchMapping("/{boardId}")
     public ResponseEntity<Response<Void>> updateNotice(@PathVariable long boardId, @Valid @RequestBody CommunityUpdateRequest request) {
 
-        noticeService.updateNotice(boardId, request);
+        noticeCommandService.updateNotice(boardId, request);
 
         return ResponseEntity.ok(Response.<Void>builder()
                 .message("공지사항 수정 성공")
@@ -46,7 +46,7 @@ public class NoticeCommandController {
     @DeleteMapping("/{boardId}")
     public ResponseEntity<Response<Void>> deleteNotice(@PathVariable long boardId) {
 
-        noticeService.deleteNotice(boardId);
+        noticeCommandService.deleteNotice(boardId);
 
         return ResponseEntity.ok(Response.<Void>builder()
                 .message("공지사항 삭제 성공")
