+++ SpringBoot/src/main/java/com/seungwook/ktsp/domain/board/type/community/notice/controller/NoticeQueryController.java
@@ -5,7 +5,7 @@ import com.seungwook.ktsp.domain.board.type.community.common.dto.response.Commun
 import com.seungwook.ktsp.domain.board.type.community.common.dto.response.CommunityResponse;
 import com.seungwook.ktsp.domain.board.type.community.common.mapper.CommunityMapper;
 import com.seungwook.ktsp.domain.board.type.community.notice.entity.Notice;
-import com.seungwook.ktsp.domain.board.type.community.notice.service.NoticeService;
+import com.seungwook.ktsp.domain.board.type.community.notice.service.NoticeQueryService;
 import com.seungwook.ktsp.domain.file.dto.AttachedFileInfo;
 import com.seungwook.ktsp.domain.file.service.FileService;
 import com.seungwook.ktsp.domain.user.mapper.UserMapper;
@@ -28,7 +28,7 @@ import java.util.List;
 public class NoticeQueryController {
 
     private final UserQueryService userQueryService;
-    private final NoticeService noticeService;
+    private final NoticeQueryService noticeQueryService;
     private final FileService fileService;
 
     @Operation(summary = "모든 공지사항 조회", description = "모든 공지사항 조회, 페이징 크기 = 15")
@@ -36,7 +36,7 @@ public class NoticeQueryController {
     public ResponseEntity<Response<Page<CommunityListResponse>>> viewNoticeList(@RequestParam(defaultValue = "0") int page) {
 
         // 모든 공지사항 조회
-        Page<CommunityListResponse> response = CommunityMapper.toNoticeList(noticeService.getAllNotice(page));
+        Page<CommunityListResponse> response = CommunityMapper.toNoticeList(noticeQueryService.getAllNotice(page));
 
         return ResponseEntity.ok(Response.<Page<CommunityListResponse>>builder()
                 .message("모든 공지사항 조회 성공")
@@ -49,7 +49,7 @@ public class NoticeQueryController {
     public ResponseEntity<Response<CommunityResponse>> viewNotice(@PathVariable long boardId) {
 
         // 공지사항
-        Notice notice = noticeService.getNotice(boardId);
+        Notice notice = noticeQueryService.getNotice(boardId);
 
         // 작성자
         Writer writer = UserMapper.toWriter(userQueryService.getWriterInfo(notice.getUser().getId()));
